fix(tasks): add button to open tasks
