+++ packages/sanity/src/structure/panes/document/statusBar/DocumentStatusBar.tsx
@@ -18,7 +18,7 @@ const CONTAINER_BREAKPOINT = 480 // px
 
 export function DocumentStatusBar(props: DocumentStatusBarProps) {
   const {actionsBoxRef} = props
-  const {editState, timelineStore, __internal_tasks} = useDocumentPane()
+  const {editState, timelineStore} = useDocumentPane()
 
   // Subscribe to external timeline state changes
   const showingRevision = useTimelineSelector(timelineStore, (state) => state.onOlderRevision)
@@ -60,7 +60,6 @@ export function DocumentStatusBar(props: DocumentStatusBarProps) {
             style={{flexShrink: 0, marginLeft: 'auto'}}
           >
             <SpacerButton size="large" />
-            {__internal_tasks && __internal_tasks.footerAction}
             {showingRevision ? <HistoryStatusBarActions /> : <DocumentStatusBarActions />}
           </Flex>
         </Flex>
